Split dynamic_api/serializers.py into a package
The file was more then 1300 lines long, and had both factory and runtime
serializer logic in the same file. This gives a more readable overview
by splitting the compile-time factory from the runtime serializer classes.
