+++ src/dso_api/dynamic_api/serializers/__init__.py
@@ -1,0 +1,8 @@
+from .base import DynamicSerializer
+from .factories import clear_serializer_factory_cache, serializer_factory
+
+__all__ = (
+    "DynamicSerializer",
+    "clear_serializer_factory_cache",
+    "serializer_factory",
+)
