+++ src/dso_api/dynamic_api/serializers/factories.py
@@ -1,17 +1,27 @@
-"""The data serialization for dynamic models.
+"""Serializer factories
 
-The serializers in this package build on top of the :mod:`rest_framework_dso.serializers`,
-to integrate the dynamic construction on top of the DSO serializer format.
-Other application-specifice logic is also introduced here as well,
-e.g. logic that depends on Amsterdam Schema policies. Such logic is *not* implemented
-in the DSO base classes as those classes are completely generic.
+This module constructs the serializer *classes* for each dataset.
+Note this doesn't do anything with the *request* object as the whole purpose
+of this module is to construct the complete serializer "field-tree" once at startup.
+
+The main function is :func:`serializer_factory`, and everything else starts from there.
+The `_links` object for instance, is backed by another serializer class, and so
+are the individual elements in that section. These elements are all serializer objects.
+
+By making everything a serializer object (instead of fields that return random dicts),
+the OpenAPI logic can generate a complete definition of the expected output.
+This allows clients (e.g. dataportaal) to generate a lovely JavaScript client based on
+the OpenAPI specification.
+
+When a request is received, these serializer class is instantiated into
+an object (in the view). The serializer *object* will also strip some fields
+that the user doesn't have access to. DRF supports that by making a deepcopy
+of the declared fields, so per-request changes don't affect the static class.
 """
 from __future__ import annotations
 
 import logging
-from functools import wraps
-from typing import Any, Callable, Iterable, Optional, TypeVar, Union, cast
-from urllib.parse import urlencode
+from typing import Optional, TypeVar, Union, cast
 
 from cachetools import LRUCache, cached
 from cachetools.keys import hashkey
@@ -20,16 +30,10 @@ from django.db import models
 from django.db.models.fields import AutoFieldMixin
 from django.db.models.fields.related import RelatedField
 from django.db.models.fields.reverse_related import ForeignObjectRel
-from django.utils.functional import SimpleLazyObject, cached_property
-from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import extend_schema_field, inline_serializer
+from django.utils.functional import SimpleLazyObject
 from more_itertools import first
 from rest_framework import serializers
-from rest_framework.exceptions import ParseError
 from rest_framework.relations import HyperlinkedRelatedField
-from rest_framework.reverse import reverse
-from rest_framework.serializers import Field
-from rest_framework.utils.model_meta import RelationInfo
 from schematools.contrib.django.factories import is_dangling_model
 from schematools.contrib.django.models import (
     DynamicModel,
@@ -37,593 +41,19 @@ from schematools.contrib.django.models import (
     LooseRelationManyToManyField,
 )
 from schematools.contrib.django.signals import dynamic_models_removed
-from schematools.types import DatasetTableSchema, Json, Temporal
+from schematools.types import DatasetTableSchema, Temporal
 from schematools.utils import to_snake_case, toCamelCase
 
-from dso_api.dynamic_api.fields import (
-    AzureBlobFileField,
-    HALLooseM2MUrlField,
-    HALLooseRelationUrlField,
-    TemporalHyperlinkedRelatedField,
-    TemporalReadOnlyField,
-)
-from dso_api.dynamic_api.permissions import filter_unauthorized_expands
-from dso_api.dynamic_api.temporal import (
-    TemporalTableQuery,
-    filter_temporal_m2m_slice,
-    filter_temporal_slice,
-)
-from dso_api.dynamic_api.utils import (
-    get_serializer_source_fields,
-    get_source_model_fields,
-    resolve_model_lookup,
-)
-from rest_framework_dso.embedding import EmbeddedFieldMatch
-from rest_framework_dso.fields import (
-    AbstractEmbeddedField,
-    DSORelatedLinkField,
-    FieldsToDisplay,
-    get_embedded_field_class,
-)
-from rest_framework_dso.serializers import (
-    DSOModelListSerializer,
-    DSOModelSerializer,
-    HALLooseLinkSerializer,
-)
-from rest_framework_dso.utils import get_serializer_relation_lookups
+from dso_api.dynamic_api.utils import get_view_name
+from rest_framework_dso.fields import AbstractEmbeddedField, get_embedded_field_class
+from rest_framework_dso.serializers import HALLooseLinkSerializer
+
+from . import base, fields
 
 MAX_EMBED_NESTING_LEVEL = 10
-logger = logging.getLogger(__name__)
-
 S = TypeVar("S", bound=serializers.Serializer)
 
-
-@extend_schema_field(
-    # Tell what this field will generate as object structure
-    inline_serializer(
-        "RelatedSummary",
-        fields={
-            "count": serializers.IntegerField(),
-            "href": serializers.URLField(),
-        },
-    )
-)
-class RelatedSummaryField(Field):
-    def to_representation(self, value: models.Manager):
-        request = self.context["request"]
-        url = reverse(get_view_name(value.model, "list"), request=request)
-
-        # the "core_filters" attribute is available on all related managers
-        filter_field = next(iter(value.core_filters.keys()))
-        q_params = {toCamelCase(filter_field + "_id"): value.instance.pk}
-
-        # If this is a temporal table, only return the appropriate records.
-        if value.model.is_temporal():
-            # The essence of filter_temporal_slice()
-            query = TemporalTableQuery(request, value.model.table_schema())
-            q_params.update(query.url_parameters)
-            value = query.filter_queryset(value.all())
-
-        query_string = ("&" if "?" in url else "?") + urlencode(q_params)
-        return {"count": value.count(), "href": f"{url}{query_string}"}
-
-
-class DynamicListSerializer(DSOModelListSerializer):
-    """This serializer class is used internally when :class:`DynamicSerializer`
-    is initialized with the ``many=True`` init kwarg to process a list of objects.
-    """
-
-    @cached_property
-    def _m2m_through_fields(self) -> tuple[models.ForeignKey, models.ForeignKey]:
-        """Give both foreignkeys that were part of this though table.
-        The fields are ordered from the perspective of this serializer;
-        with the first one being the field that was used to enter the through table,
-        and the second to jump in the target relation.
-        """
-        target_model = self.child.Meta.model
-        if not target_model.table_schema().through_table:
-            raise RuntimeError("not an m2m table")
-
-        first_fk = get_source_model_fields(self.parent, self.field_name, self)[-1].remote_field
-        second_fk = first(
-            (f for f in target_model._meta.get_fields() if f.is_relation and f is not first_fk)
-        )
-        return first_fk, second_fk
-
-    def get_attribute(self, instance: DynamicModel):
-        """Override list attribute retrieval for temporal filtering.
-        The ``get_attribute`` call is made by DRF to retrieve the data from an instance,
-        based on our ``source_attrs``.
-        """
-        value = super().get_attribute(instance)
-
-        # When the source attribute of this object references the reverse_name field
-        # (from a ForeignKey), the value is a RelatedManager subclass.
-        if isinstance(value, models.Manager):
-            request = self.context["request"]
-            if value.model.is_temporal():
-                # Make sure the target table is filtered by the temporal query,
-                # instead of showing links to all records.
-                value = filter_temporal_slice(request, value.all())
-
-            if (
-                value.model.table_schema().through_table
-                and self._m2m_through_fields[1].related_model.is_temporal()
-            ):
-                # When this field lists the through table entries, also filter the target table.
-                # Otherwise, the '_links' section shows all references from the M2M table, while
-                # the embedding returns fewer entries because they are filtered temporal query.
-                value = filter_temporal_m2m_slice(
-                    request, value.all(), self._m2m_through_fields[1]
-                )
-
-        return value
-
-    def get_queryset_iterator(self, queryset: models.QuerySet) -> Iterable[models.Model]:
-        """Optimize querysets to access only the fields our serializer needs.
-
-        This checks which model fields "self.fields" will access, and limits the
-        queryset accordingly. In effect, this reduces field access when "?_fields=..." is used
-        on the request, and it avoids requesting database fields that the user has no access to.
-        """
-        only_fields = get_serializer_source_fields(self)
-        if only_fields:
-            # Make sure no unnecessary fields are requested from the database.
-            # This avoids accessing privileged data, and reduces bandwidth for ?_fields=.. queries.
-            queryset = queryset.only(*only_fields)
-
-        return super().get_queryset_iterator(queryset)
-
-    def get_prefetch_lookups(self) -> list[Union[models.Prefetch, str]]:
-        """Optimize prefetch lookups
-
-        - Return correct temporal records only (for M2M tables).
-        - Make sure prefetches only read the subset of user-accessible fields.
-        """
-        relation_lookups = get_serializer_relation_lookups(self)
-
-        lookups = []
-        for lookup, field in relation_lookups.items():
-            if lookup.startswith("rev_m2m_"):
-                # The reverse m2m lookups don't seem to benefit from prefetch related:
-                continue
-
-            # The prefetch object get a custom queryset, so unwanted rows/fields are excluded.
-            prefetch_queryset = self._get_prefetch_queryset(lookup, field)
-            lookups.append(models.Prefetch(lookup, queryset=prefetch_queryset))
-
-        return lookups
-
-    def _get_prefetch_queryset(self, lookup: str, field: serializers.Field) -> models.QuerySet:
-        """Construct a queryset for a prefetch on the ORM path lookup path.
-        The serializer field is passed as well, to detect which source fields it might need.
-        """
-        # Find which model fields the relation would traverse.
-        # This information is not available on serializer field objects (nor can't be)
-        parent_model = self.child.Meta.model
-        model_field = resolve_model_lookup(parent_model, lookup)[-1]
-
-        # Avoid prefetching unnecessary fields from the database.
-        only_fields = self._get_prefetch_fields(field, model_field)
-        prefetch_queryset = model_field.related_model.objects.only(*only_fields)
-
-        # Avoid prefetching unnecessary temporal records for M2M relations
-        # (that is, records that not part of the current temporal slice).
-        if cast(DynamicModel, model_field.related_model).is_temporal() and (
-            # Tell whether the field returns many results (e.g. M2M).
-            model_field.many_to_many
-            or model_field.one_to_many
-        ):
-            request = self.context["request"]
-            prefetch_queryset = filter_temporal_slice(request, prefetch_queryset)
-
-        return prefetch_queryset
-
-    def _get_prefetch_fields(
-        self,
-        field: serializers.Field,
-        model_field: RelatedField | ForeignObjectRel | LooseRelationField,
-    ) -> list[str]:
-        """Tell which fields the prefetch query should retrieve."""
-        if isinstance(field, serializers.BaseSerializer):
-            only_fields = get_serializer_source_fields(field)
-            if isinstance(model_field, models.ManyToOneRel):  # A reverse FK (ManyToOneRel)
-                # For reverse relations, the reverse foreign key needs to be provided too.
-                # Otherwise, our call to prefetch_related_objects() fails on deferred attributes.
-                only_fields.append(model_field.field.attname)
-            return only_fields
-        else:
-            raise NotImplementedError(f"Can't determine reduced fields subset yet for: {field}")
-
-    @cached_property
-    def expanded_fields(self) -> list[EmbeddedFieldMatch]:
-        """Filter unauthorized fields from the matched expands."""
-        return filter_unauthorized_expands(
-            self.context["request"].user_scopes,
-            expanded_fields=super().expanded_fields,
-            skip_unauth=self.expand_scope.auto_expand_all,
-        )
-
-
-class DynamicSerializer(DSOModelSerializer):
-    """The logic for dynamically generated serializers.
-
-    Most DSO-related logic is found in the base class
-    :class:`~rest_framework_dso.serializers.DSOModelSerializer`.
-    This class adds more application-specific logic to the serializer such as:
-
-    * Logic that depends on Amsterdam Schema data.
-    * Permissions based on the schema.
-    * Temporal relations (ideally in DSO, but highly dependent on schema data).
-    * The ``_links`` section logic (depends on knowledge of temporal relations).
-
-    A part of the serializer construction happens here
-    (the Django model field -> serializer field translation).
-    The remaining part happens in :func:`serializer_factory`.
-    """
-
-    _default_list_serializer_class = DynamicListSerializer
-
-    serializer_field_mapping = {
-        **DSOModelSerializer.serializer_field_mapping,
-        LooseRelationField: HALLooseRelationUrlField,
-        LooseRelationManyToManyField: HALLooseRelationUrlField,
-    }
-
-    schema = serializers.SerializerMethodField()
-
-    table_schema: DatasetTableSchema = None
-
-    @cached_property
-    def _request(self):
-        """Get request from this or parent instance."""
-        # Avoids resolving self.root all the time:
-        return self.context["request"]
-
-    def get_fields(self) -> dict[str, serializers.Field]:
-        """Override DRF to remove fields that shouldn't be part of the response."""
-        fields = {}
-        for field_name, field in super().get_fields().items():
-            if self._apply_permission(field_name, field):
-                fields[field_name] = field
-
-        return fields
-
-    def _apply_permission(self, field_name: str, field: serializers.Field) -> bool:
-        """Check permissions and apply any transforms."""
-        user_scopes = self._request.user_scopes
-
-        if field.source == "*":
-            # e.g. _links field, always include. These sub serializers
-            # do their own permission checks for their fields.
-            return True
-
-        # Find which ORM path is traversed for a field.
-        # (typically one field, except when field.source has a dotted notation)
-        model_fields = get_source_model_fields(self, field_name, field)
-        for model_field in model_fields:
-            field_schema = DynamicModel.get_field_schema(model_field)
-
-            # Check access
-            permission = user_scopes.has_field_access(field_schema)
-            if not permission:
-                return False
-
-            # Check transform from permission
-            if transform_function := permission.transform_function():
-                # Value must be transformed, decorate to_representation() for it.
-                # Fields are a deepcopy, so this doesn't affect other serializer instances.
-                # This strategy also avoids having to dig into the response data afterwards.
-                field.to_representation = self._apply_transform(
-                    field.to_representation, transform_function
-                )
-
-        return True
-
-    @staticmethod
-    def _apply_transform(
-        to_representation: Callable[[Any], Json], transform_function: Callable[[Json], Json]
-    ):
-        """Make sure an additional transformation happens before the field generated output.
-        :param to_representation: The bound method from the serializer
-        :param transform_function: A function that adjusts the rendered value.
-        """
-
-        @wraps(to_representation)
-        def _new_to_representation(data):
-            value = to_representation(data)
-            return transform_function(value)
-
-        return _new_to_representation
-
-    @cached_property
-    def expanded_fields(self) -> list[EmbeddedFieldMatch]:
-        """Filter unauthorized fields from the matched expands."""
-        return filter_unauthorized_expands(
-            self._request.user_scopes,
-            expanded_fields=super().expanded_fields,
-            skip_unauth=self.expand_scope.auto_expand_all,
-        )
-
-    @classmethod
-    def get_embedded_field(cls, field_name, prefix="") -> AbstractEmbeddedField:
-        """Overridden to improve error messages.
-        At this layer, information about the Amsterdam schema
-        can be used to provide a better error message.
-        """
-        try:
-            return super().get_embedded_field(field_name, prefix=prefix)
-        except ParseError:
-            # Raise a better message, or fallback to the original message
-            cls._raise_better_embed_error(field_name)
-            raise
-
-    @classmethod
-    def _raise_better_embed_error(cls, field_name, prefix=""):
-        """Improve the error message for embedded fields."""
-        # Check whether there is a RelatedSummaryField for a link,
-        # to tell that this field doesn't get expanded.
-        try:
-            link_field = cls._declared_fields["_links"]._declared_fields[field_name]
-        except KeyError:
-            return
-        else:
-            if isinstance(link_field, RelatedSummaryField):
-                raise ParseError(
-                    f"The field '{prefix}{field_name}' is not available"
-                    f" for embedding as it's a summary of a huge listing."
-                ) from None
-
-    def get_embedded_objects_by_id(
-        self, embedded_field: AbstractEmbeddedField, id_list: list[Union[str, int]]
-    ) -> Union[models.QuerySet, Iterable[models.Model]]:
-        """Retrieve a number of embedded objects by their identifier.
-
-        This override makes sure the correct temporal slice is returned.
-        """
-        if embedded_field.is_loose:
-            # Loose relations always point to a temporal object. In this case, the link happens on
-            # the first key only, and the temporal slice makes sure that a second WHERE condition
-            # happens on a temporal field (volgnummer/beginGeldigheid/..).
-            # When 'identifier' is ["identificatie", "volgnummer"], take the first as grouper.
-            model = embedded_field.related_model
-            id_field = embedded_field.related_id_field or model.table_schema().identifier[0]
-            return filter_temporal_slice(self._request, model.objects.all()).filter(
-                **{f"{id_field}__in": id_list}
-            )
-        else:
-            queryset = super().get_embedded_objects_by_id(embedded_field, id_list=id_list)
-
-            # For all relations: if they are in fact temporal objects
-            # still only return those that fit within the current timeframe.
-            return filter_temporal_slice(self._request, queryset)
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_schema(self, instance):
-        """The schema field is exposed with every record"""
-        table = instance.get_table_id()
-        dataset_path = instance.get_dataset_path()
-        return f"https://schemas.data.amsterdam.nl/datasets/{dataset_path}/dataset#{table}"
-
-    def build_url_field(self, field_name, model_class):
-        """Make the URL to 'self' is properly initialized."""
-        if issubclass(self.serializer_url_field, serializers.Serializer):
-            # Temporal serializer.
-            field_kwargs = {"source": "*"}
-        else:
-            # Normal DSOSelfLinkField, link to correct URL.
-            field_kwargs = {"view_name": get_view_name(model_class, "detail")}
-
-        return self.serializer_url_field, field_kwargs
-
-    def build_relational_field(self, field_name: str, relation_info: RelationInfo):
-        """Make sure temporal links get a different field class."""
-        field_class, field_kwargs = super().build_relational_field(field_name, relation_info)
-        related_model = relation_info.related_model
-
-        if "view_name" in field_kwargs:
-            # Fix the view name to point to our views.
-            field_kwargs["view_name"] = get_view_name(related_model, "detail")
-
-        # Upgrade the field type when it's a link to a temporal model.
-        if field_class is DSORelatedLinkField and related_model.table_schema().is_temporal:
-            # Ideally this would just be an upgrade to a different field class.
-            # However, since the "identificatie" and "volgnummer" fields are dynamic,
-            # a serializer class is better suited as field type. That ensures the sub object
-            # is properly generated in the OpenAPI spec as a distinct class type.
-            field_class = _temporal_link_serializer_factory(relation_info.related_model)
-            field_kwargs.pop("queryset", None)
-            field_kwargs.pop("view_name", None)
-
-        return field_class, field_kwargs
-
-    def build_property_field(self, field_name, model_class):
-        """This is called for the foreignkey_id fields.
-        As the field name doesn't reference the model field directly,
-        DRF assumes it's an "@property" on the model.
-        """
-        model_field = model_class._meta._forward_fields_map.get(field_name)
-        if (
-            model_field is not None
-            and isinstance(model_field, models.ForeignKey)
-            and model_field.related_model.is_temporal()
-        ):
-            return TemporalReadOnlyField, {}
-        else:
-            return super().build_property_field(field_name, model_class)
-
-
-class DynamicBodySerializer(DynamicSerializer):
-    """This subclass of the dynamic serializer only exposes the non-relational fields.
-
-    Ideally, this should be obsolete as the serializer_factory() can avoid
-    generating those fields in the first place.
-    """
-
-    def get_fields(self):
-        """Remove fields that shouldn't be in the body."""
-        fields = super().get_fields()
-
-        # Remove fields from the _links field too. This is not done in the serializer
-        # itself as that creates a cross-dependency between the parent/child.fields property.
-        links_field = fields.get("_links")
-        if links_field is not None and isinstance(links_field, DynamicLinksSerializer):
-            if self.fields_to_display.reduced():
-                # The 'invalid_fields' is not checked against here, as that already happened
-                # for the top-level fields reduction.
-                main_and_links_fields = self.get_valid_field_names(fields)
-                fields_to_keep, _ = self.fields_to_display.get_allow_list(main_and_links_fields)
-                fields_to_keep.update(links_field.fields_always_included)
-                links_field.fields = {
-                    name: field
-                    for name, field in links_field.fields.items()
-                    if name in fields_to_keep
-                }
-
-        return fields
-
-    def get_valid_field_names(self, fields: dict[str, serializers.Field]) -> set[str]:
-        """Override to allow limiting the ``_links`` field layout too.
-        By allowing the names from the ``_links`` field as if these exist at this level,
-        those fields can be restricted quite naturally. This is needed to unclutter
-        the ``_links`` output when nested relations are queried.
-        """
-        valid_names = super().get_valid_field_names(fields)
-
-        # When the _links object is a serializer, also allow these fields
-        # names as if these are present here.
-        links_field = fields.get("_links")
-        if links_field is not None and isinstance(links_field, DynamicLinksSerializer):
-            links_field.parent = self  # perform early Field.bind()
-            links_fields = set(links_field.fields.keys()) - links_field.fields_always_included
-            valid_names.update(links_fields)
-
-        return valid_names
-
-
-class DynamicLinksSerializer(DynamicSerializer):
-    """The serializer for the ``_links`` field.
-    Following DSO/HAL guidelines, it contains "self", "schema", and all relational fields.
-    """
-
-    fields_always_included = {"self", "schema"}
-
-    @property
-    def fields_to_display(self):
-        # Make sure child serializers (e.g. temporal relations / through serializers)
-        # don't reduce their fields by taking this object from their "parent" serializer.
-        return FieldsToDisplay()
-
-    @property
-    def expanded_fields(self) -> list[EmbeddedFieldMatch]:
-        # No need for our super class try look for expanded fields.
-        return super().expanded_fields
-
-    def limit_return_fields(self, fields):
-        # Don't apply the ?_fields=... request here, let the parent to this instead.
-        # Otherwise there is an initialization loop in get_fields() to find
-        # the fields of the parent and child at the same time.
-        return fields
-
-    def get_fields(self) -> dict[str, serializers.Field]:
-        """Override to avoid adding fields that loop back to a parent."""
-        fields = super().get_fields()
-
-        if self.parent_source_fields:
-            for name, field in fields.copy().items():
-                if self._is_repeated_relation(name, field):
-                    if logger.isEnabledFor(logging.DEBUG):
-                        logger.debug(
-                            "Excluded %s from _links field, it resolves to a parent via %s.%s",
-                            name,
-                            ".".join(f.name for f in self.parent_source_fields),
-                            field.source,
-                        )
-                    del fields[name]
-
-        return fields
-
-    @cached_property
-    def parent_source_fields(
-        self,
-    ) -> list[Union[RelatedField, LooseRelationField, ForeignObjectRel]]:
-        """Find the ORM relationship fields that lead to this serializer instance"""
-        source_fields = []
-        parent = self.parent
-
-        # The root-level obviously doesn't have a field name,
-        # hence the check for parent.parent to exclude the root object.
-        while parent.parent is not None:
-            if parent.source_attrs and hasattr(parent.parent, "Meta"):
-                model = parent.parent.Meta.model
-
-                for attr in parent.source_attrs:
-                    field = model._meta.get_field(attr)
-                    source_fields.append(field)
-                    model = field.related_model
-
-            parent = parent.parent
-        return source_fields
-
-    def _is_repeated_relation(self, name: str, field: serializers.Field) -> bool:
-        """Check whether the field references back to a parent relation.
-        This detects whether the field is a relation, and whether it's model fields
-        are references by the pre-calculated `source_fields`.
-        """
-        if field.source == "*" or not isinstance(
-            field,
-            (
-                serializers.ModelSerializer,
-                serializers.ListSerializer,  # e.g. list of ModelSerializer
-                serializers.RelatedField,
-                serializers.ManyRelatedField,
-            ),
-        ):
-            return False
-
-        # Resolve the model field that this serializer field links to.
-        # The source_attrs is generated here as it's created later in field.bind().
-        model_fields = get_source_model_fields(self, name, field)
-        for model_field in model_fields:
-            # Check whether the reverse relation of that model field
-            # is already walked through by a parent serializer
-            if (
-                model_field.remote_field is not None
-                and model_field.remote_field in self.parent_source_fields
-            ):
-                return True
-
-        return False
-
-    def _include_embedded(self):
-        """Never include embedded relations when the
-        serializer is used for generating the _links section.
-        """
-        return False
-
-
-def get_view_name(model: type[DynamicModel], suffix: str):
-    """Return the URL pattern for a dynamically generated model.
-
-    :param model: The dynamic model.
-    :param suffix: This can be "detail" or "list".
-    """
-    dataset_id = to_snake_case(model.get_dataset_id())
-    table_id = to_snake_case(model.get_table_id())
-    return f"dynamic_api:{dataset_id}-{table_id}-{suffix}"
-
-
-class ThroughSerializer(DynamicSerializer):
-    """The serializer for the through table of M2M relations.
-    It's fields are defined by the factory function, so little code is found here.
-    """
-
-    def limit_return_fields(self, fields):
-        # Don't limit a through serializer via ?_fields=...,
-        # as this exists in the _links section.
-        return fields
-
-
+logger = logging.getLogger(__name__)
 _serializer_factory_cache = LRUCache(maxsize=100000)
 _temporal_link_serializer_factory_cache = LRUCache(maxsize=100000)
 
@@ -711,12 +141,27 @@ def _serializer_cache_key(model, depth=0, nesting_level=0):
 @cached(cache=_serializer_factory_cache, key=_serializer_cache_key)
 def serializer_factory(
     model: type[DynamicModel], depth: int = 0, nesting_level=0
-) -> type[DynamicSerializer]:
+) -> type[base.DynamicSerializer]:
     """Generate the DRF serializer class for a specific dataset model.
 
     Internally, this creates all serializer fields based on the metadata
     from the model, and underlying schema definition. It also generates
     a secondary serializer for the ``_links`` field where all relations are exposed.
+
+    In short, this factory does in-memory would you'd typically write as:
+
+    .. code-block:: python
+
+        class SomeDatasetTableSerializer(DynamicSerializer):
+            _links = SomeDatasetTableLinksSerializer(source="*")
+            field1 = serializers.CharField(...)
+            field2 = serializers.WhateverField(...)
+
+            class Meta:
+                model = SomeDatasetTable
+                fields = ("_links", "field1", "field2")
+
+    ...and all child fields are statically defined on the instance whenever possible.
 
     Following DSO/HAL guidelines, objects are serialized with ``_links`` and ``_embedded``
     fields, but without relational fields at the top level. The relational fields appear
@@ -797,7 +242,7 @@ def serializer_factory(
     _generate_nested_relations(serializer_part, model, nesting_level)
 
     # Generate Meta section and serializer class
-    return serializer_part.construct_class(serializer_name, base_class=DynamicBodySerializer)
+    return serializer_part.construct_class(serializer_name, base_class=base.DynamicBodySerializer)
 
 
 def _validate_model(model: type[DynamicModel]):
@@ -812,7 +257,7 @@ def _validate_model(model: type[DynamicModel]):
 
 def _links_serializer_factory(
     model: type[DynamicModel], depth: int
-) -> type[DynamicLinksSerializer]:
+) -> type[base.DynamicLinksSerializer]:
     """Generate the DRF serializer class for the ``_links`` section."""
     safe_dataset_id = to_snake_case(model.get_dataset_id())
     serializer_name = f"{safe_dataset_id.title()}{model.__name__}LinksSerializer"
@@ -843,12 +288,12 @@ def _links_serializer_factory(
             _build_serializer_reverse_fk_field(serializer_part, model_field)
         # This includes LooseRelationManyToManyField
         elif isinstance(model_field, models.ManyToManyField):
-            _build_m2m_serializer_field(serializer_part, model, model_field)
+            _build_m2m_serializer_field(serializer_part, model_field)
         elif isinstance(model_field, (RelatedField, ForeignObjectRel, LooseRelationField)):
             _build_serializer_links_field(serializer_part, model_field)
 
     # Generate serializer class
-    return serializer_part.construct_class(serializer_name, base_class=DynamicLinksSerializer)
+    return serializer_part.construct_class(serializer_name, base_class=base.DynamicLinksSerializer)
 
 
 def _nontemporal_link_serializer_factory(
@@ -878,7 +323,7 @@ def _nontemporal_link_serializer_factory(
     # Construct the class
     safe_dataset_id = to_snake_case(related_model.get_dataset_id())
     serializer_name = f"{safe_dataset_id.title()}{related_model.__name__}LinkSerializer"
-    return serializer_part.construct_class(serializer_name, serializers.ModelSerializer)
+    return serializer_part.construct_class(serializer_name, base_class=serializers.ModelSerializer)
 
 
 @cached(cache=_temporal_link_serializer_factory_cache)
@@ -913,7 +358,9 @@ def _temporal_link_serializer_factory(
     serializer_part.add_field(
         "href",
         _build_href_field(
-            related_model, field_cls=TemporalHyperlinkedRelatedField, table_schema=table_schema
+            related_model,
+            field_cls=fields.TemporalHyperlinkedRelatedField,
+            table_schema=table_schema,
         ),
     )
     if related_model.has_display_field():
@@ -932,7 +379,7 @@ def _temporal_link_serializer_factory(
     # Construct the class
     safe_dataset_id = to_snake_case(related_model.get_dataset_id())
     serializer_name = f"{safe_dataset_id.title()}{related_model.__name__}LinkSerializer"
-    return serializer_part.construct_class(serializer_name, serializers.ModelSerializer)
+    return serializer_part.construct_class(serializer_name, base_class=serializers.ModelSerializer)
 
 
 def _loose_link_serializer_factory(
@@ -953,7 +400,7 @@ def _loose_link_serializer_factory(
     )
     serializer_part.add_field(
         "href",
-        _build_href_field(related_model, field_cls=HALLooseRelationUrlField),
+        _build_href_field(related_model, field_cls=fields.HALLooseRelationUrlField),
     )
     if related_model.has_display_field():
         serializer_part.add_field("title", serializers.CharField(source="*"))
@@ -967,7 +414,7 @@ def _loose_link_serializer_factory(
     serializer_name = f"{safe_dataset_id.title()}{related_model.__name__}LinkSerializer"
     serializer_part.class_attrs.pop("Meta")  # we dont need Meta on regular Serializers
 
-    return serializer_part.construct_class(serializer_name, HALLooseLinkSerializer)
+    return serializer_part.construct_class(serializer_name, base_class=HALLooseLinkSerializer)
 
 
 def _build_serializer_field(  # noqa: C901
@@ -1037,7 +484,7 @@ def _build_serializer_embedded_field(
     )
 
     embedded_field = EmbeddedFieldClass(
-        serializer_class=cast(type[DynamicSerializer], serializer_class),
+        serializer_class=cast(type[base.DynamicSerializer], serializer_class),
         # serializer_class=serializer_class,
         source=model_field.name,
     )
@@ -1050,7 +497,7 @@ def _build_serializer_embedded_field(
 
 def _through_serializer_factory(  # noqa: C901
     m2m_field: models.ManyToManyField,
-) -> type[ThroughSerializer]:
+) -> type[base.ThroughSerializer]:
     """Generate the DRF serializer class for a M2M model.
 
     This works directly on the database fields of the through model,
@@ -1098,9 +545,9 @@ def _through_serializer_factory(  # noqa: C901
     href_field_cls = HyperlinkedRelatedField
     field_kwargs = {}
     if loose:
-        href_field_cls = HALLooseM2MUrlField
+        href_field_cls = fields.HALLooseM2MUrlField
     elif temporal is not None:
-        href_field_cls = TemporalHyperlinkedRelatedField
+        href_field_cls = fields.TemporalHyperlinkedRelatedField
         field_kwargs["table_schema"] = target_table_schema
     serializer_part.add_field(
         "href",
@@ -1160,13 +607,11 @@ def _through_serializer_factory(  # noqa: C901
     # Finalize as serializer
     safe_dataset_id = to_snake_case(through_model.get_dataset_id())
     serializer_name = f"{safe_dataset_id.title()}{toCamelCase(through_model.__name__)}_M2M"
-    return serializer_part.construct_class(serializer_name, base_class=ThroughSerializer)
+    return serializer_part.construct_class(serializer_name, base_class=base.ThroughSerializer)
 
 
 def _build_m2m_serializer_field(
-    serializer_part: SerializerAssemblyLine,
-    model: type[DynamicModel],
-    m2m_field: models.ManyToManyField,
+    serializer_part: SerializerAssemblyLine, m2m_field: models.ManyToManyField
 ):
     """Add a serializer for a m2m field to the output parameters.
 
@@ -1247,7 +692,7 @@ def _build_serializer_blob_field(
     camel_name = toCamelCase(model_field.name)
     serializer_part.add_field(
         camel_name,
-        AzureBlobFileField(
+        fields.AzureBlobFileField(
             account_name=field_schema["account_name"],
             source=(model_field.name if model_field.name != camel_name else None),
         ),
@@ -1281,7 +726,7 @@ def _build_serializer_reverse_fk_field(
         serializer_part.add_field(name, field_class(read_only=True, many=True))
     elif format1 == "summary":
         # Only shows a count and href to the (potentially large) list of items.
-        serializer_part.add_field(name, RelatedSummaryField())
+        serializer_part.add_field(name, fields.RelatedSummaryField())
     else:
         logger.warning("Field %r uses unsupported format: %s", field_schema, format1)
